+++ scancodeio/__init__.py
@@ -30,7 +30,7 @@ from django.conf import settings
 # Django starts so that shared_task will use this app.
 from scancodeio.celery import app as celery_app
 
-__version__ = "1.0.5"
+__version__ = "1.0.6"
 
 SCAN_NOTICE = Path(__file__).resolve().parent.joinpath("scan.NOTICE").read_text()